section_14/pyproject.toml
[tool.poetry]
name = "qrcode_projeto"
version = "0.1.0"
description = ""
authors = ["José Matheus R. Flora <jmrfloraxbox@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
customtkinter = "^5.2.1"
pillow = "^10.1.0"
packaging = "^23.2"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
